FIX: Lower case addressable in Explode string
We were lowercasing the addressable when the index was still part of the token. This meant for an category specific parameters. i.e. instantmort.m{MALE} would get changed to instantmort.m{male}. If your category labels were MALE and not male then Casal2 would error out. This is fixed.