UPD: initialisation phase
MVE: the report execute for kInitialisation from model.cpp to the initialisation manger so we can print reports after each initialsation phase.
CHG: child method from Execute -> DoExecute. This allowed me to pass the model pointer to the Execute(class) so that the manager could use the model to execute the reports.
ADD: a check if we have scaled the partition for B0 models. For multiple initialisation phases we were re-scaling each phase which was an error.